+++ common/src/main/java/net/enderitemc/enderitemod/misc/EnderiteSulfurCubeRenderState.java
@@ -1,0 +1,7 @@
+package net.enderitemc.enderitemod.misc;
+
+public interface EnderiteSulfurCubeRenderState {
+    void enderitemod$setCrackedEnderiteOre(boolean crackedEnderiteOre);
+
+    boolean enderitemod$isCrackedEnderiteOre();
+}