+++ fabric/src/main/java/net/enderitemc/enderitemod/mixin/fabric/EnderiteCrossbowHeldItemRendererMixin.java
@@ -40,7 +40,7 @@ public abstract class EnderiteCrossbowHeldItemRendererMixin {
     protected abstract void applyItemArmTransform(PoseStack matrices, HumanoidArm arm, float equipProgress);
 
     @Inject(
-        method = "renderArmWithItem(Lnet/minecraft/client/player/AbstractClientPlayer;FFLnet/minecraft/world/InteractionHand;FLnet/minecraft/world/item/ItemStack;FLcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/SubmitNodeCollector;I)V",
+        method = "submitArmWithItem(Lnet/minecraft/client/player/AbstractClientPlayer;FFLnet/minecraft/world/InteractionHand;FLnet/minecraft/world/item/ItemStack;FLcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/SubmitNodeCollector;I)V",
         at = @At(value = "INVOKE", target = "Lnet/minecraft/world/item/ItemStack;isEmpty()Z"),
         cancellable = true
     )
@@ -147,7 +147,7 @@ public abstract class EnderiteCrossbowHeldItemRendererMixin {
 
     @WrapOperation(
         at = @At(value = "INVOKE", target = "Lnet/minecraft/world/item/ItemStack;getUseDuration(Lnet/minecraft/world/entity/LivingEntity;)I"),
-        method = "renderArmWithItem(Lnet/minecraft/client/player/AbstractClientPlayer;FFLnet/minecraft/world/InteractionHand;FLnet/minecraft/world/item/ItemStack;FLcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/SubmitNodeCollector;I)V")
+        method = "submitArmWithItem(Lnet/minecraft/client/player/AbstractClientPlayer;FFLnet/minecraft/world/InteractionHand;FLnet/minecraft/world/item/ItemStack;FLcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/SubmitNodeCollector;I)V")
     private int enderitemod$changeBowTime(ItemStack instance, LivingEntity user, Operation<Integer> original) {
         if (instance.is(EnderiteTools.ENDERITE_BOW.get())) {
             int maxTime = original.call(instance, user);